requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-dotenv>=1.0.1
passlib==1.7.4
bcrypt==3.2.2
